remove tripInterstitialContent component, rework into registrationForm
