+++ src/contexts/FormContext.js
@@ -8,9 +8,9 @@ export default function FormProvider({ children }) {
   useEffect(() => {
     if (travelersData.length !== 0) {
       // TODO: ADAPT-4677 - Determine how to cross-reference the travelers data with their associated addresses
-      console.log('Prefill form data: ', travelersData);
+      console.log('Prefill form data: ', { ...travelersData });
       // TODO: ADAPT-4677 - Update prefill obj with travelers AND their associated addresses
-      window.prefill = travelersData;
+      window.prefill = { ...travelersData };
     }
   }, [travelersData]);
